chore(deps): bump go deps
