+++ f3-sidecar/utils.go
@@ -2,12 +2,15 @@ package main
 
 import (
 	"context"
+	"fmt"
+	"os"
 	"path/filepath"
 	"time"
 
 	"github.com/filecoin-project/go-f3/gpbft"
 	"github.com/ipfs/go-cid"
 	leveldb "github.com/ipfs/go-ds-leveldb"
+	logging "github.com/ipfs/go-log/v2"
 )
 
 var CID_UNDEF_RUST = cid.MustParse("baeaaaaa")
@@ -43,4 +46,39 @@ func getNetworkName(rawNetworkName string) gpbft.NetworkName {
 		networkName = "filecoin"
 	}
 	return networkName
+}
+
+func setLogLevel(name string, level string) error {
+	if err := logging.SetLogLevel(name, level); err != nil {
+		return fmt.Errorf("%s %w", name, err)
+	}
+	return nil
+}
+
+func setLogLevels() error {
+	logging.SetAllLoggers(logging.LevelInfo)
+	if err := setLogLevel("dht", "error"); err != nil {
+		return err
+	}
+	// Always mute RtRefreshManager because it breaks terminals
+	if err := setLogLevel("dht/RtRefreshManager", "warn"); err != nil {
+		return err
+	}
+	if err := setLogLevel("f3/sidecar", "debug"); err != nil {
+		return err
+	}
+	return nil
+}
+
+func checkError(err error) {
+	if err != nil {
+		panic(err)
+	}
+}
+
+// To avoid potential panics
+// See <https://github.com/ChainSafe/forest/pull/4636#issuecomment-2306500753>
+func setGoDebugEnv() {
+	err := os.Setenv("GODEBUG", "invalidptr=0,cgocheck=0")
+	checkError(err)
 }
