+++ go.work
@@ -1,4 +1,4 @@
-go 1.24.5
+go 1.25.2
 
 use (
 	./f3-sidecar
